requirements.txt
paho-mqtt
requests
google-auth